fix: rename env to settings
