+++ src/slices/session.ts
@@ -1,6 +1,6 @@
 import Cookies from "js-cookie"
 
-import { SESSION_METADATA_COOKIE_NAME } from "../env"
+import { SESSION_METADATA_COOKIE_NAME } from "../settings"
 import createSlice from "./createSlice"
 
 export interface SessionState {
